requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
python-barcode==0.15.1
pyzbar==0.1.9
